API

The following pages document and explain the functionality of all exported types and functions in the library.

Index

SpectralStatistics.BerryRobnik

SpectralStatistics.BerryRobnikBrody

SpectralStatistics.Brody

SpectralStatistics.DataSample

SpectralStatistics.GOE

SpectralStatistics.GSE

SpectralStatistics.GUE

SpectralStatistics.Gamma

SpectralStatistics.Model

SpectralStatistics.Poisson

SpectralStatistics.RealSpectrum

SpectralStatistics.UnfoldedSpectrum

SpectralStatistics.level_spacing

SpectralStatistics.level_spacing_cdf

SpectralStatistics.level_spacing_pdf

SpectralStatistics.level_spacing_u

SpectralStatistics.number_variance

SpectralStatistics.rigidity

SpectralStatistics.spectral_form_factor

SpectralStatistics.unfold_spectrum

Abstract types

SpectralStatistics.DataSample — Type

DataSample

DataSample is an abstract supertype encompassing all concrete implementations of data structures used for processing spectra in the SpectralStatistics.jl library.

SpectralStatistics.Model — Type

Model

Model is an abstract supertype encompassing all concrete implementations of the analytical spectral statistics models in the SpectralStatistics.jl library.

Data types

SpectralStatistics.RealSpectrum — Type

RealSpectrum <: DataSample

RealSpectrum is a concrete type used as a container for spectra of real numbers.

SpectralStatistics.UnfoldedSpectrum — Type

UnfoldedSpectrum <: DataSample

UnfoldedSpectrum is a concrete type used as a container for spectra after unfolding.

Spectral Statistics

SpectralStatistics.level_spacing — Function

level_spacing(spect::UnfoldedSpectrum; n::Int = 1) → s::Vector

Return the level spacings of order n. The nearest neighbour level spacings are the default, given by n=1.

Description

The level spacings of order n are given by the difference between the n-th consecutive levels of the spectrum. The nearest neighbour level spacing is considered most commonly.

Arguments

spect: The unfolded energy spectrum, given as an instance of type UnfoldedSpectrum.

Keyword arguments

n=1 : The order of the level spacings.

Returns

s : Vector of level spacings.

SpectralStatistics.level_spacing_pdf — Function

level_spacing_pdf(model::Model, pts::Vector; n::Int = 1) → p::Vector

Return the analytical expression for the level spacing probability density function, corresponding to the chosen model, evaluated at positions pts.

Description

The nearest neighbour level spacing distributions are the most commonly studied spectral statistics.

Arguments

model: The model, given as an instance of a concrete subtype of Model.

pts: The positions where probabability density function should be evaluated.

Keyword arguments

n=1 : The order of the level spacings.

Returns

p : Vector of the probabilites.

level_spacing_pdf(spect::UnfoldedSpectrum, bins::Vector; n::Int = 1) → s::Vector p::Vector

Return a histogram of the probability density function of the level spacings of order n. The nearest neighbour level spacings are the default, given by n=1.

Description

The nearest neighbour level spacing distributions are the most commonly studied spectral statistics.

Arguments

spect: The unfolded energy spectrum, given as an instance of type UnfoldedSpectrum.

bins: The boundaries of the bin positions.

Keyword arguments

n=1 : The order of the level spacings.

Returns

p : Vector of the probability contained in each bin.

SpectralStatistics.level_spacing_cdf — Function

level_spacing_cdf(model::Model, pts::Vector; n::Int = 1) → w::Vector

Return the analytical expression for level spacing cumulative density function, corresponding to the chosen model, evaluated at positions pts.

Arguments

model: The model, given as an instance of a concrete subtype of Model.

pts: The positions where cumulative density function should be evaluated.

Keyword arguments

n=1 : The order of the level spacings.

Returns

w : Vector of the cumulative probabilities.

level_spacing_cdf(spect::UnfoldedSpectrum, pts::Vector; n::Int = 1) → s::Vector w::Vector

Return the cumulative density function of the level spacings of order n evaluated at positions pts. The nearest neighbour level spacings are the default, given by n=1.

Arguments

spect: The unfolded energy spectrum, given as an instance of type UnfoldedSpectrum.

pts: The positions where cumulative density function should be evaluated.

Keyword arguments

n=1 : The order of the level spacings.

Returns

w : Vector of the cumulative probabilities.

SpectralStatistics.level_spacing_u — Function

level_spacing_u(model::Model, pts::Vector; n::Int = 1) → u::Vector

Return the analytical expression for the spectraly normalized cumulative density function of the nearest neighbour level spacings evaluated at positions pts.

Arguments

model: The model, given as an instance of a concrete subtype of Model.

pts: The positions where cumulative density function should be evaluated.

Keyword arguments

n=1 : The order of the level spacings.

Returns

u : Vector of the cumulative probabilities.

level_spacing_u(spect::UnfoldedSpectrum, pts::Vector; n::Int = 1) → s::Vector u::Vector

Return the spectraly normalized cumulative density function of the nearest neighbour level spacings evaluated at positions pts.

Description

The nearest neighbour level spacing distributions are the most commonly studied spectral statistics. To normalize the relative fluctuations it is useful to perform the following nonlinear transformation

\[U(s) := \frac{2}{\pi}\arccos\sqrt{1-W(s)},\]

where $W(s)$ is the cumulative level spacing distribution.

Arguments

spect: The unfolded energy spectrum, given as an instance of type UnfoldedSpectrum.

pts: The positions where cumulative density function should be evaluated.

Keyword arguments

n=1 : The order of the level spacings.

Returns

u : Vector of the cumulative probabilities.

SpectralStatistics.number_variance — Function

number_variance(model::Model, l) → p::Vector

Number variance.

number_variance(spect::S, x::Vector{T}) → s::Vector

Number variance.

SpectralStatistics.rigidity — Function

rigidity(model::Model, l::Vector) → p::Vector

Spectral rigidity.

SpectralStatistics.spectral_form_factor — Function

spectral_form_factor(model::Model, t) → p::Vector

Spectral form factor.

spectral_form_factor(E::Vector{T}, t::T) → s::Vector

Spectral form factor.

Analytical models

SpectralStatistics.Poisson — Type

Poisson <: Model

Poisson is a concrete type used to represent the Poisson model.

Description

The Poisson model applies to sequences of independent random variables. Based on the Berry-Tabor conjecture the spectral statistics (in the semiclassical limit) of integrable systems are described by this model.

Attributes

This model has no attributes.

API

The following spectral statistcs can be evaluated for this model:

level_spacing_pdf

level_spacing_cdf

level_spacing_u

number_variance

rigidity

spectral_form_factor

SpectralStatistics.GOE — Type

GOE <: Model

GOE is a concrete type used to represent the Gaussian Orthogonal Ensemble model of random matrix theory.

Description

Based on the Bohigas-Giannoni-Schmit conjecture this model describes the spectral statistics (in the semiclassical limit) of chaotic systems, invariant under time-reversal symmetry and without spin 1/2 interactions.

Attributes

This model has no attributes.

API

The following spectral statistcs can be evaluated for this model:

level_spacing_pdf

level_spacing_cdf

level_spacing_u

number_variance

rigidity

spectral_form_factor

SpectralStatistics.GUE — Type

GUE <: Model

GUE is a concrete type used to represent the Gaussian Unitary Ensemble model of random matrix theory.

Description

Based on the Bohigas-Giannoni-Schmit conjecture this model describes the spectral statistics (in the semiclassical limit) of chaotic systems, without time-reversal (or a generalized anti-unitary) symmetry.

Attributes

This model has no attributes.

API

The following spectral statistcs can be evaluated for this model:

level_spacing_pdf

level_spacing_cdf

level_spacing_u

number_variance

rigidity

spectral_form_factor

SpectralStatistics.GSE — Type

GSE <: Model

GSE is a concrete type used to represent the Gaussian Symplectic Ensemble model of random matrix theory.

Description

Based on the Bohigas-Giannoni-Schmit conjecture this model describes the spectral statistics (in the semiclassical limit) of chaotic systems, invariant under time-reversal symmetry and including spin 1/2 interactions.

Attributes

This model has no attributes.

API

The following spectral statistcs can be evaluated for this model:

level_spacing_pdf

level_spacing_cdf

level_spacing_u

number_variance

rigidity

spectral_form_factor

SpectralStatistics.Gamma — Type

Gamma <: Model

Gamma is a concrete type used to represent the Gamma model.

Description

This model represents the gamma distributions.

Attributes

gamma: The level repulsion exponent.

API

The following spectral statistcs can be evaluated for this model:

level_spacing_pdf

level_spacing_cdf

SpectralStatistics.Brody — Type

Brody <: Model

Brody is a concrete type used to represent the Brody model.

Description

This model interpolates between the Poisson ($\beta=0$) and Wigner-Dyson ($\beta=1$) distributions. It is commonly used to describe systems with some degree of localization.

Attributes

beta: The level repulsion exponent.

API

The following spectral statistcs can be evaluated for this model:

level_spacing_pdf

level_spacing_cdf

level_spacing_u

SpectralStatistics.BerryRobnik — Type

BerryRobnik <: Model

BerryRobnik is a concrete type used to represent the Berry-Robnik model (with two components).

Description

This model describes the spectral statistics (in the semiclassical limit) of systems whose classical phase space features both regular and chaotic motion. These systems are commonly refered to as systems with divided phase space or mixed-type systems. The model rests on the argument that in the semiclassical limit the spectrum subspectra belonging to states that describe belonging to the states to the distinct types of motion (regular and chaotic) will decompose into separate components. The regular part of the spectrum is modeled by Poisson statistics and the chaotic part by GOE statistics.

Attributes

rho: The Liouville measure of the combined regular component.

API

The following spectral statistcs can be evaluated for this model:

level_spacing_pdf

level_spacing_cdf

level_spacing_u

number_variance

SpectralStatistics.BerryRobnikBrody — Type

BerryRobnikBrody <: Model

BerryRobnikBrody is a concrete type used to represent the Berry-Robnik-Brody model.

Description

This model is the Berry-Robnik model (see BerryRobnik) where instead of the GOE the Brody model is used for the chaotic part. It is commonly used to describe systems with a divided phase space as well as some degree of localization.

Attributes

rho: The Liouville measure of the combined regular component.

beta: The level repulsion exponent.

API

The following spectral statistcs can be evaluated for this model:

level_spacing_pdf

level_spacing_cdf

level_spacing_u

Functions

SpectralStatistics.unfold_spectrum — Function

unfold_spectrum(spect::DataSample, f::Function) → unfolded::UnfoldedSpectrum

Return the unfolded spectrum of spect by using the function f as the smooth part of the integrated density of states.

Arguments

spect: Spectrum to be unfolded.

f(x) : Function modeling the smooth part of the integrated density of states, where the argument x is the energy.

Returns

unfolded : The unfolded spectrum as an instance of type UnfoldedSpectrum.

unfold_spectrum(spect::DataSample, n::Int) → unfolded::UnfoldedSpectrum

Return the unfolded spectrum of spect by fitting a polynomial of degree n to the integrated density of states.

Arguments

n : Degree of polynomial moddeling the smooth part of the integrated density of states.

unfold_spectrum(spect::DataSample, n::Int, cut_values) → unfolded::UnfoldedSpectrum

Return the unfolded spectrum of spect by piecewise fitting polynomials of degree n to the integrated density of states.

Arguments

cut_values : Relative positions of the cuts between the spectral pieces.